타워 업그레이드 알고리즘 수정
타워 업그레이드 에셋 추가,
다음 할일 각 타워 만들기

+++ Defense/Assets/2. Scripts/PlayerState.cs
@@ -5,8 +5,8 @@ using UnityEngine.UI;
 
 public class PlayerState : MonoBehaviour
 {
-    int startMoney = 300;
-    public int currentMoney;
+    float startMoney = 300f;
+    public float currentMoney;
     public int currentLife;
     int maxlife = 3;
